refactor(mcp-server): move evals file into the clients/mcp-server folder and remove the modifications to root package.json
